Consortium tests working except for skipped error cases, which may need rewriting
